More moving classes around.
