+++ src/main/scala/io/archivesunleashed/matchbox/ExtractGraph.scala
@@ -16,10 +16,9 @@
  */
 package io.archivesunleashed.matchbox
 
-import io.archivesunleashed.ArchiveRecord
-import io.archivesunleashed.Transformations._
-import io.archivesunleashed.matchbox.StringUtils._
-import io.archivesunleashed.utils.JsonUtil
+import io.archivesunleashed._
+import matchbox.StringUtils._
+import utils.JsonUtil
 
 import org.apache.spark.graphx._
 import org.apache.spark.rdd.RDD
